+++ BE/src/main/java/com/codegenie/workbook/repository/WorkbookRepository.java
@@ -3,4 +3,4 @@ package com.codegenie.workbook.repository;
 import com.codegenie.workbook.entity.Workbook;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface WorkbookRepository extends JpaRepository<Workbook, Long> { }
+public interface WorkbookRepository extends JpaRepository<Workbook, Integer> { }
